Add files via upload
Home page partially done.
I've everything commented explaining how it was done and the CSS is commented aswell so just copy and paste everything from that if you're working on a page. All the images we should use are there including the brand image and favicon which should be on all the pages.